refactor: test 코드 정리
